# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Restful way to use Palvelutietovaranto.",
    "version" : "v1",
    "title" : "RESTful PTV"
  },
  "basePath" : "/",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "paths" : {
    "/electronicServiceChannels" : {
      "get" : {
        "tags" : [ "ElectronicServiceChannels", "ServiceChannels" ],
        "summary" : "Lists electronic service channels",
        "description" : "Lists electronic service channels",
        "operationId" : "listElectronicServiceChannels",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns array of electronic service channels",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ElectronicServiceChannel"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/electronicServiceChannels/{electronicServiceChannelId}" : {
      "get" : {
        "tags" : [ "ElectronicServiceChannels", "ServiceChannels" ],
        "summary" : "finds an electronic service channel by id",
        "description" : "finds an electronic service channel by id",
        "operationId" : "findElectronicServiceChannel",
        "parameters" : [ {
          "name" : "electronicServiceChannelId",
          "in" : "path",
          "description" : "electronic service channel id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an electronic service channel",
            "schema" : {
              "$ref" : "#/definitions/ElectronicServiceChannel"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/services/{serviceId}" : {
      "get" : {
        "tags" : [ "Services" ],
        "summary" : "Finds a service by id",
        "description" : "Finds service",
        "operationId" : "findService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Services" ],
        "summary" : "Updates service",
        "description" : "Updates service",
        "operationId" : "updateService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Organization service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Service"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an updated service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "WebPage" : {
      "type" : "object",
      "required" : [ "language", "type", "url", "value" ],
      "properties" : {
        "description" : {
          "type" : "string",
          "description" : "Web page description."
        },
        "url" : {
          "type" : "string",
          "description" : "Web page url.",
          "maxLength" : 500
        },
        "language" : {
          "type" : "string",
          "description" : "Language code."
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      },
      "description" : "Open API web page model."
    },
    "LocalizedListItem" : {
      "type" : "object",
      "required" : [ "language", "type", "value" ],
      "properties" : {
        "language" : {
          "type" : "string",
          "description" : "Language code."
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "LanguageItem" : {
      "type" : "object",
      "required" : [ "language", "value" ],
      "properties" : {
        "value" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Email" : {
      "type" : "object",
      "properties" : {
        "value" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Phone" : {
      "type" : "object",
      "properties" : {
        "additionalInformation" : {
          "type" : "string"
        },
        "serviceChargeType" : {
          "type" : "string"
        },
        "chargeDescription" : {
          "type" : "string"
        },
        "prefixNumber" : {
          "type" : "string"
        },
        "isFinnishServiceNumber" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Service" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "type" : {
          "type" : "string"
        },
        "statutoryDescriptionId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "serviceClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FintoItem"
          }
        },
        "ontologyTerms" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FintoItem"
          }
        },
        "targetGroups" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FintoItem"
          }
        },
        "lifeEvents" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FintoItem"
          }
        },
        "industrialClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FintoItem"
          }
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedListItem"
          }
        },
        "descriptions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedListItem"
          }
        },
        "languages" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "pattern" : "^[a-z]{2}$"
          }
        },
        "keywords" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "coverageType" : {
          "type" : "string"
        },
        "municipalities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Municipality"
          }
        },
        "webPages" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "requirements" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LanguageItem"
          }
        },
        "publishingStatus" : {
          "type" : "string"
        },
        "chargeType" : {
          "type" : "string"
        },
        "additionalInformations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedListItem"
          }
        },
        "organizationIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "electronicServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "phoneServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "printableFormServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "serviceLocationServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "webPageServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      }
    },
    "FintoItem" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "name" : {
          "type" : "string"
        },
        "code" : {
          "type" : "string"
        },
        "ontologyType" : {
          "type" : "string"
        },
        "uri" : {
          "type" : "string"
        },
        "parentId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "parentUri" : {
          "type" : "string"
        }
      }
    },
...
    "ElectronicServiceChannel" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "type" : {
          "type" : "string"
        },
        "organizationId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedListItem"
          }
        },
        "descriptions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedListItem"
          }
        },
        "signatureQuantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "requiresSignature" : {
          "type" : "boolean"
        },
        "supportPhones" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Phone"
          }
        },
        "supportEmails" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "requiresAuthentication" : {
          "type" : "boolean"
        },
        "urls" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LanguageItem"
          }
        },
        "languages" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "attachments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Attachment"
          }
        },
        "webPages" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "serviceHours" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceHour"
          }
        },
        "publishingStatus" : {
          "type" : "string"
        }
      }
    },
...
    "Attachment" : {
      "type" : "object",
      "required" : [ "language", "url" ],
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string",
          "maxLength" : 100
        },
        "description" : {
          "type" : "string",
          "maxLength" : 150
        },
        "url" : {
          "type" : "string",
          "maxLength" : 500
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "ServiceHour" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string",
          "description" : "Type of service hour (Standard, Exception or Special)."
        },
        "exceptionHourType" : {
          "type" : "string",
          "description" : "Type of service hour exception type. Valid values are: Open or Closed."
        },
        "validFrom" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time where from this entry is valid."
        },
        "validTo" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time to this entry is valid."
        },
        "monday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on monday."
        },
        "tuesday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on tuesday."
        },
        "wednesday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on wednesday."
        },
        "thursday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on thursday."
        },
        "friday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on friday."
        },
        "saturday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on saturday."
        },
        "sunday" : {
          "type" : "boolean",
          "description" : "Is this service hour effective on sunday."
        },
        "opens" : {
          "type" : "string",
          "description" : "Opening time in format HH:mm for example 08:00."
        },
        "closes" : {
          "type" : "string",
          "description" : "Closing time in format HH:mm for example 19:00"
        },
        "additionalInformation" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LanguageItem"
          }
        }
      }
    },
    "Municipality" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string",
          "description" : "Municipality code (like 491 or 091)."
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LanguageItem"
          }
        }
      }
    },
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "NotFound" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
  }
}